Parser.py: auto-detect result values
Now that we can override tests result values with a config file, it is more
convenient to auto-detect the test result values rather than hardcoding.

This changes a bit the summary table and group sections order and fields,
but is still arguably acceptable.

While at it, print a one line summary during execution.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>